requirements.txt
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
numpy==2.2.6
requests==2.32.4
urllib3==2.5.0
flask==3.0.0
flask-cors==4.0.0
bcrypt==4.0.1
firebase_admin
google-cloud-firestore
python-jose
pdfplumber
python-docx
chromadb
langchain
PyMuPDF
python-docx
sentence_transformers
ollama
openai
ftfy
spacy
textstat
sentence_transformers
